Full repository import #11689

+++ modules/lib/lib-export/src/main/java/com/enonic/xp/lib/export/ImportHandler.java
@@ -30,11 +30,11 @@ public class ImportHandler
 
     private Boolean includePermissions;
 
-    private Function<Long, Void> nodeImported;
+    private Function<Integer, Void> nodeImported;
 
-    private Function<Long, Void> nodeResolved;
+    private Function<Integer, Void> nodeResolved;
 
-    private Function<Long, Void> nodeSkipped;
+    private Function<Integer, Void> nodeSkipped;
 
     public NodeImportResultMapper execute()
     {
@@ -106,17 +106,17 @@ public class ImportHandler
         this.includePermissions = includePermissions;
     }
 
-    public void setNodeImported( final Function<Long, Void> nodeImported )
+    public void setNodeImported( final Function<Integer, Void> nodeImported )
     {
         this.nodeImported = nodeImported;
     }
 
-    public void setNodeResolved( final Function<Long, Void> nodeResolved )
+    public void setNodeResolved( final Function<Integer, Void> nodeResolved )
     {
         this.nodeResolved = nodeResolved;
     }
 
-    public void setNodeSkipped( final Function<Long, Void> nodeSkipped )
+    public void setNodeSkipped( final Function<Integer, Void> nodeSkipped )
     {
         this.nodeSkipped = nodeSkipped;
     }
